重构调色器窗口逻辑
